feat(blog): What's New on home hero and account menu
Track unread release posts in localStorage, surface summaries on the hero
updates row, and add account dropdown entry without forcing dynamic layout.
